+++ packages/brownfield-navigation/src/scripts/brownfield.ts
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import { runCLI } from '@callstack/brownfield-cli';
+
+runCLI(process.argv);