+++ packages/brownie/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @callstack/brownie
 
+## 3.2.1
+
+### Patch Changes
+
+- [#265](https://github.com/callstack/react-native-brownfield/pull/265) [`7e114e6`](https://github.com/callstack/react-native-brownfield/commit/7e114e676fc290a54912b4d99bd22952b2f7380d) Thanks [@hurali97](https://github.com/hurali97)! - version bump
+
+- Updated dependencies [[`7e114e6`](https://github.com/callstack/react-native-brownfield/commit/7e114e676fc290a54912b4d99bd22952b2f7380d)]:
+  - @callstack/brownfield-cli@3.2.1
+
+## 3.2.0
+
+### Minor Changes
+
+- [#263](https://github.com/callstack/react-native-brownfield/pull/263) [`b7dfa7e`](https://github.com/callstack/react-native-brownfield/commit/b7dfa7e30e1a921bb7ac0952e23481ee081fb9ed) Thanks [@hurali97](https://github.com/hurali97)! - version bump for release
+
+### Patch Changes
+
+- Updated dependencies [[`b7dfa7e`](https://github.com/callstack/react-native-brownfield/commit/b7dfa7e30e1a921bb7ac0952e23481ee081fb9ed)]:
+  - @callstack/brownfield-cli@3.2.0
+
+## 3.1.0
+
+### Minor Changes
+
+- [#236](https://github.com/callstack/react-native-brownfield/pull/236) [`3e33b89`](https://github.com/callstack/react-native-brownfield/commit/3e33b890e6f647f6bbebe32f8068a38b1ed85ea0) Thanks [@hurali97](https://github.com/hurali97)! - add brownfield navigation
+
+### Patch Changes
+
+- Updated dependencies [[`3e33b89`](https://github.com/callstack/react-native-brownfield/commit/3e33b890e6f647f6bbebe32f8068a38b1ed85ea0)]:
+  - @callstack/brownfield-cli@3.1.0
+
 ## 3.0.0
 
 ### Patch Changes